refactored into individual pages and started creating routes to save data
