+++ package.json
@@ -16,6 +16,7 @@
     "@types/react-dom": "18.2.7",
     "@vercel/postgres": "^0.4.1",
     "autoprefixer": "10.4.15",
+    "axios": "^1.5.0",
     "dotenv": "^16.3.1",
     "eslint": "8.47.0",
     "eslint-config-next": "13.4.18",
