+++ src/app/page.tsx
@@ -1,23 +1,19 @@
 "use client";
 
 import React, { useState } from "react";
+import Link from "next/link";
 import prisma from "../../lib/prisma";
-import Resume from "./components/resume";
 import Form from "./components/resumeForm/resumeForm";
 
 export default function Home() {
-  const [toggleInputs, setToggleInputs] = useState(true);
-
   return (
     <div className="m-auto bg-white w-[80vw] h-[90vh] border p-4">
-      <button
-        className="bg-blue-400 hover:bg-blue-300 text-white font-bold py-2 px-4 rounded"
-        onClick={() => setToggleInputs(!toggleInputs)}
-      >
-        {toggleInputs ? "Show Resume" : "Show Form"}
-      </button>
-
-      {toggleInputs ? <Form /> : <Resume />}
+      <Link href="/form" className="bg-blue-400 hover:bg-blue-300 text-white font-bold py-2 px-4 rounded mr-4">
+        Show Form
+      </Link>
+      <Link href="/resume" className="bg-green-400 hover:bg-green-300 text-white font-bold py-2 px-4 rounded">
+        Show Resume
+      </Link>
     </div>
   );
 }
